requirements.txt
PyMuPDF
pytesseract
Pillow
pdf2image
langchain
langchain-community
langchain-core
ollama
sentence-transformers
faiss-cpu
langchain
streamlit
langchain_groq
torch==2.1.0
sentence-transformers==2.2.2
